Renames chamber access to sql access.
Chamber access was unnecessarily ambiguous.
